+++ index.js
@@ -1,5 +1,6 @@
 const http = require('http');
 const express = require('express');
+
 const app = express();
 
 app.use(express.static('public'));
@@ -9,4 +10,3 @@ const server = http.createServer(app);
 server.listen(app.get('port'), () => {
   console.log(`Server listening on port ${app.get('port')}...`);
 });
-
