+++ public/js/logic.js
@@ -1,0 +1,9 @@
+import { rotateChar } from './utils/util.js';
+
+const convertRot13 = function(sourceText) {
+  return sourceText.split('').map(rotateChar).join('');
+};
+
+export {
+  convertRot13,
+};
